requirements.txt
streamlit==1.29.0
tensorflow==2.15.0
pillow==10.1.0
matplotlib==3.8.2
numpy==1.26.3
scikit-learn==1.3.2
